Delegates added. Still working on it though

+++ LooseChange/DTO/LooseChangeDTO.cpp
@@ -4,7 +4,7 @@ LooseChangeDTO::LooseChangeDTO()
 {
 }
 
-LooseChangeDTO::LooseChangeDTO(int inId, QDate inDate, float inAmount, TransactionType inTransactionType, Category inCategory, QString inComment)
+LooseChangeDTO::LooseChangeDTO(int inId, QDate inDate, double inAmount, TransactionType inTransactionType, Category inCategory, QString inComment)
 {
     id = inId;
     date = inDate;
